Did some restructuring of the repo, so function calls work properly
