+++ generate_iso_template.py
@@ -3,7 +3,7 @@
 import argparse
 from argparse import RawTextHelpFormatter
 import sys
-from asf_metadata import iso_template2lists, add_dem_lists, \
+from metadata.asf_metadata import iso_template2lists, add_dem_lists, \
   iso_xml_structure, meta_xml_file, iso_dictionary_structure, meta_json_file
