+++ metadata/util.py
